pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning[plugin]"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
metadata = true
style = "semver"

[tool.poetry]
name = "luke_bot"
version = "0.0.0"
authors = [
  "Exlot <author@example.com>" , "Charlie Wilson <charles.ep.wilson@gmail.com>",
]
description = "Discord bot for posting Luke updates"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Framework :: Pytest",
    "Natural Language :: English",
    "Topic :: Games/Entertainment",
]

[tool.poetry.dependencies]
python = "^3.11"

python-dotenv = "^1.0.0"
"discord.py" = "^2.0.1"
pydantic = "^1.10"
startgg = "^0.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
ruff = "^0.0.262"
black = "^23.3.0"
docformatter = { extras = ["tomli"], version = "^1.6.2" }
pyright = "^1.1.304"
requests = "^2.28.1"
types-requests = "^2.28.6"

[tool.poetry.urls]
"Homepage" = "https://github.com/LukeBotFanClub/LukeBot"
"repository" = "https://github.com/LukeBotFanClub/LukeBot.git"
"Bug Tracker" = "https://github.com/LukeBotFanClub/LukeBot/issues"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.ruff]
# Enable flake8-bugbear (`B`) rules.
select = ["E", "F", "B", "I"]

# Never enforce `E501` (line length violations).
ignore = ["E501"]

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]

[tool.black]
target-version = ['py311']
preview = true

[tool.docformatter]
recursive = true

[tool.pyright]
pythonVersion = "3.11"
